requirements.txt
# FastAPI core
fastapi==0.115.5
starlette==0.40.0
uvicorn==0.32.0

# Pydantic v2 (new standard)
pydantic==2.9.2
pydantic-core==2.23.4
typing-extensions>=4.12.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
itsdangerous==2.2.0  # lightweight token helper for forms if needed
slowapi==0.1.9       # rate limiting still compatible

# Firebase / Firestore
firebase-admin==6.5.0
google-cloud-firestore==2.18.0
google-auth==2.35.0
google-api-core==2.19.1

# Scheduling & Utilities
apscheduler==3.10.4
bleach==6.1.0
email-validator==2.2.0

# Optional / recommended
httpx==0.27.2        # newer async HTTP client (replaces requests)
python-dotenv==1.0.1 # environment variable management
